+++ app/src/main/kotlin/com/hosshan/android/kodic/store/UserStore.kt
@@ -1,8 +1,9 @@
 package com.hosshan.android.kodic.store
 
 import android.content.Context
-import com.hosshan.android.kodic.CodicService
+import com.hosshan.android.kodic.CodicEngineService
 import com.hosshan.android.kodic.model.Project
+import com.hosshan.android.kodic.store.service.CodicUserService
 import com.hosshan.android.kodic.util.CodicRetrofitUtil
 import rx.Observable
 import kotlin.platform.platformStatic
@@ -10,20 +11,12 @@ import kotlin.platform.platformStatic
 /**
  * Created by shunhosaka on 15/09/12.
  */
-public class UserStore {
+public class UserStore(val service: CodicUserService) {
 
-    companion object {
-        @JvmStatic fun getProject(context: Context, id: Int): Observable<Project?>
-                = CodicRetrofitUtil
-                .buildRestAdapter(context)
-                .create(javaClass<CodicService>())
-                .getProject(id)
+    fun getProject(id: Int): Observable<Project?> =
+            service.getProject(id)
 
-        @JvmStatic fun getProjectList(context: Context): Observable<List<Project>>
-                = CodicRetrofitUtil
-                .buildRestAdapter(context)
-                .create(javaClass<CodicService>())
-                .getProjectList()
-    }
+    fun getProjectList(context: Context): Observable<List<Project>> =
+            service.getProjectList()
 
 }
