+++ app/src/main/kotlin/com/hosshan/android/kodic/store/service/CodicEngineService.kt
@@ -13,7 +13,7 @@ import rx.Observable
 /**
  * Created by a40875 on 2015/09/10.
  */
-public interface CodicService {
+public interface CodicEngineService {
 
     @GET("/v1/engine/translate.json")
     fun getTranslate(@Query("text") text: String, @Query("project_id") projectId: Int, @Query("casing") casing: String, @Query("acronym_style") acronymStyle: String): Observable<List<TranslatedText>>
@@ -24,15 +24,5 @@ public interface CodicService {
     @POST("/v1/engine/translate.json")
     fun postTranslate(): Observable<List<TranslatedText>>
 
-    @GET("/v1/user_projects.json")
-    fun getProjectList(): Observable<List<Project>>
 
-    @GET("/v1/user_projects/{id}.json")
-    fun getProject(@Path("id") id: Int): Observable<Project?>
-
-    @GET("/v1/ced/lookup.json")
-    fun getCedLookup(@Query("query") query: String, @Query("count") number: Int = 10): Observable<List<CedTitle>>
-
-    @GET("/v1/ced/entries/{id}.json")
-    fun getCedEntry(@Path("id") id: Int): Observable<CedEntry>
 }
